fix: check exit codes of running `gh` (#991)
We used to not check the success of calling subprocesses. That means that sometimes our secrets did and sometimes didn't propagate, without us knowing.

This is probably the reason of some broken passwords in the past...

(To be clear: if we get throttled this change won't solve the issue. But at least we'll know about it!)