fix: command no longer retries on throttling (#1014)
In https://github.com/cdklabs/aws-secrets-github-sync/pull/1003 the
retry loop was broken because we stopped capturing `stderr`.

Because we didn't capture `stderr` we could no longer inspect the
error output for the error message, and so we didn't retry (in
fact if the underlying `gh` tool failed, so would we).

Re-capture `stderr`, and print the underlying tool output to the screen
(but only if an input is provided in the case of `stdout`... this was
added without explaining motivation in the original PR, so I kept the
behavior in order to not mess anything up)